fixed dm and dr variance issues
